raiden: transport: move _handle_closed_connection to WebRTCManager
